Set up V2 of the API endpoints

+++ app/features/companies/companies.controller.ts
@@ -2,6 +2,7 @@ import { ICompany } from "@/app/lib/interface/ICompany"
 import { Result } from "@/app/types"
 import { NextRequest, NextResponse } from "next/server"
 import * as companiesService from "./companies.service"
+import mongoose from "mongoose"
 
 const getCompanyData = (data: unknown) => {
   return data &&
@@ -59,8 +60,8 @@ export const createCompany = async (
 }
 
 export const listCompanies = async (filter?: {
-  search: string
-  companyName: string
+  search: string | null
+  companyName: string | null
 }): Promise<NextResponse<Result<ICompany[]>>> => {
   const companies = await companiesService.list(filter)
   if (!companies.success)
@@ -78,6 +79,17 @@ export const listCompanies = async (filter?: {
 export const getCompanyById = async (
   companyId: string,
 ): Promise<NextResponse<Result<ICompany>>> => {
+  if (!mongoose.isValidObjectId(companyId)) {
+    return NextResponse.json(
+      {
+        success: false,
+        error:
+          "Invalid id format. Must be must be a string of 12 bytes or a string of 24 hex characters or a number.",
+      },
+      { status: 400 },
+    )
+  }
+
   const company = await companiesService.single(companyId)
   if (!company.success) {
     switch (company.type) {
@@ -124,6 +136,16 @@ export const updateCompanyById = async (
       { status: 415 },
     )
   }
+  if (!mongoose.isValidObjectId(companyId)) {
+    return NextResponse.json(
+      {
+        success: false,
+        error:
+          "Invalid id format. Must be must be a string of 12 bytes or a string of 24 hex characters or a number.",
+      },
+      { status: 400 },
+    )
+  }
   const body = await req.json()
   const companyData = getCompanyData(body)
 
@@ -168,6 +190,16 @@ export const deleteCompanyById = async (
   req: NextRequest,
   companyId: string,
 ): Promise<NextResponse<Result<ICompany>>> => {
+  if (!mongoose.isValidObjectId(companyId)) {
+    return NextResponse.json(
+      {
+        success: false,
+        error:
+          "Invalid id format. Must be must be a string of 12 bytes or a string of 24 hex characters or a number.",
+      },
+      { status: 400 },
+    )
+  }
   const deletedCompanyResult = await companiesService.remove(companyId)
 
   if (!deletedCompanyResult.success) {
